requirements.txt
requests
python-solv
zstandard
